# public/api.yaml
info:
  title: Coronatime API
  description: Here goes description
  version: 0.1.9
servers:
  - url: http://localhost:3000/wane/v1
    description: local server
  - url: http://producation.app/api
    description: production server
openapi: 3.0.0
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
          description: ID du client
        surnom:
          type: string
          description: Surnom du client
        telephone:
          type: string
          description: Telephone du client
        adrresse:
          type: string
          description: Adresse du client
      required:
        - surnom
        - telephone
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /clients/telephone:
    post:
      operationId: findByTelephone
      tags:
        - Clients
      summary: "Récupérer un client par numéro de téléphone"
      description: "Cette méthode permet de récupérer un client spécifique en fonction de son numéro de téléphone."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                telephone:
                  type: string
                  description: "Numéro de téléphone du client à rechercher"
                  example: "+1234567890"
              required:
                - telephone
      responses:
        '200':
          description: "Client récupéré avec succès"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: "Client non trouvé"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Client not found"
        '400':
          description: "Bad Request - Requête invalide"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid request"
...
  /clients/{id}/dettes:
    get:
      operationId: findDettes

      summary: "Afficher les dettes d'un client spécifique"
      tags:
        - Clients
      description: "Cette méthode permet de récupérer toutes les dettes associées à un client en fonction de son ID."
      parameters:
        - name: id
          in: path
          required: true
          description: "ID du client pour lequel récupérer les dettes"
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: "Client retrouvé avec succès avec ses dettes"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientWithDettes'
        '404':
          description: "Client non trouvé"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Client non trouvé"
        '400':
          description: "Requête invalide"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Une erreur s'est produite"

    post:
      operationId: storeClient
      tags:
        - Clients
      summary: "Créer un nouveau client"
      description: "Cette méthode permet de créer un nouveau client et, éventuellement, un utilisateur associé avec une photo."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  example: "Doe"
                prenom:
                  type: string
                  example: "John"
                telephone:
                  type: string
                  example: "+1234567890"
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                user:
                  type: object
                  properties:
                    login:
                      type: string
                      example: "jdoe"
                    password:
                      type: string
                      format: password
                      example: "password123"
                photo:
                  type: string
                  format: binary
                  description: "Photo de l'utilisateur"
              required:
                - nom
                - prenom
                - telephone
                - email
      responses:
        '200':
          description: "Client créé avec succès"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientWithUser'
        '422':
          description: "Erreur de validation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation failed"
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
        '500':
          description: "Erreur interne du serveur"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Une erreur s'est produite"
...